backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["yoonki <yoonkeee@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sniffio = "^1.3.0"
fastapi = "^0.94.0"
uvicorn = "^0.21.0"
pymysql = "^1.0.2"
django-environ = "^0.10.0"
ipdb = "^0.13.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
